ClaimFlow/pyproject.toml
[project]
name = "claimflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django==5.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
    "djoser>=2.3.3",
    "python-dotenv>=1.2.1",
]
